A lot of changes.
Major ones,the way collision is checked is changed:
the collision list checks if the object says it can collide with each
other, if it can it asks how high a priority it is then uses the higher
priorities collide method.

Collidable is now a functional interface(abstract class) for non-movable
objects

Tester.java is where I'm puting starting states for testing

static lines can be added, and things will bounce off of them as one
might expect

a form of box which is just 4 static lines now exists, it can contain
moving balls.

pressing clicking makes a ball of mass 50-100. on release. dragging
before letting go causes the ball to have a velocity in that direction
proportional to to the distance dragged(1pixel/second for every pixel)

spheres(when colliding with lines/boxes/spheres travel(back in time) to
the point of collision, does the collision method then'fast forwards' to
the current time.

Spheres are randomly colored to make it easier to identify them.

Movable contains a method for reflection of normal vector and center of
mass collisions